Add files via upload
Write a java program to create an abstract class named Shape that contains an empty method
named numberOfSides( ). Provide three classes named Rectangle, Triangle and Hexagon such that
each one of the classes extends the class Shape. Each one of the classes contains only the method
numberOfSides( ) that shows the number of sides in the given geometrical structures